Progress with searching for stops, still faulty somewhere
