+++ app/src/main/AndroidManifest.xml
@@ -4,7 +4,7 @@
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
-        android:name=".di.Application"
+        android:name="com.jikokujo.core.di.Application"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
@@ -12,9 +12,10 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
+        android:usesCleartextTraffic="true"
         android:theme="@style/Theme.JikokuJó">
         <activity
-            android:name=".MainActivity"
+            android:name="com.jikokujo.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.JikokuJó">
