--- app/src/main/java/com/jikokuj/di/Application.kt
@@ -1,8 +1,0 @@
-package com.jikokuj.di
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class Application : Application() {
-}
