--- app/src/main/java/com/jikokuj/presentation/schedule/Schedule.kt
@@ -1,2 +1,0 @@
-package com.jikokuj.presentation.schedule
-
