+++ app/src/main/java/com/jikokujo/core/data/ApiResponseStructure.kt
@@ -1,0 +1,11 @@
+package com.jikokujo.core.data
+
+import com.google.gson.annotations.SerializedName
+
+interface Payload
+data class ResponseRoot<T: Payload>(
+    @SerializedName("data")
+    val data: T? = null,
+    @SerializedName("errors")
+    val errors: List<String> = listOf()
+)
