+++ app/src/main/java/com/jikokujo/core/di/Application.kt
@@ -1,0 +1,7 @@
+package com.jikokujo.core.di
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+@HiltAndroidApp
+class Application : Application() {
+}
