+++ gradle/libs.versions.toml
@@ -11,7 +11,7 @@ composeBom = "2025.11.01"
 hiltVersion = "2.57.1"
 kspVersion = "2.3.3"
 retrofitVersion = "3.0.0"
-kotlinSerialization = "1.9.0"
+gsonVersion = "2.13.2"
 mapsforgeVersion = "0.26.0"
 
 [libraries]
@@ -33,8 +33,8 @@ androidx-compose-material3 = { group = "androidx.compose.material3", name = "mat
 dagger-hilt-android = { group = "com.google.dagger", name = "hilt-android", version.ref = "hiltVersion" }
 dagger-hilt-android-compiler = { group = "com.google.dagger", name = "hilt-android-compiler", version.ref = "hiltVersion" }
 retrofit = { group = "com.squareup.retrofit2", name = "retrofit", version.ref = "retrofitVersion" }
-kotlinx-serialization = { group = "org.jetbrains.kotlinx", name = "kotlinx-serialization-json", version.ref = "kotlinSerialization"}
-kotlinx-serialization-converter = { group = "com.squareup.retrofit2", name = "converter-kotlinx-serialization", version.ref = "retrofitVersion"}
+gson-serialization = { group = "com.google.code.gson", name = "gson", version.ref = "gsonVersion" }
+gson-serialization-converter = { group = "com.squareup.retrofit2", name = "converter-gson", version.ref = "retrofitVersion"}
 mapsforge-core = { group = "com.github.mapsforge.mapsforge", name = "mapsforge-core", version.ref = "mapsforgeVersion" }
 mapsforge-map = { group = "com.github.mapsforge.mapsforge", name = "mapsforge-map", version.ref = "mapsforgeVersion" }
 mapsforge-android = { group = "com.github.mapsforge.mapsforge", name = "mapsforge-map-android", version.ref = "mapsforgeVersion" }
@@ -45,6 +45,5 @@ mapsforge-map-reader = { group = "com.github.mapsforge.mapsforge", name = "mapsf
 android-application = { id = "com.android.application", version.ref = "agp" }
 kotlin-android = { id = "org.jetbrains.kotlin.android", version.ref = "kotlin" }
 kotlin-compose = { id = "org.jetbrains.kotlin.plugin.compose", version.ref = "kotlin" }
-kotlin-serialization = { id = "org.jetbrains.kotlin.plugin.serialization", version.ref = "kotlinSerialization" }
 dagger-hilt-android = { id = "com.google.dagger.hilt.android", version.ref = "hiltVersion" }
 devtools-ksp = { id = "com.google.devtools.ksp", version.ref = "kspVersion" }
